@@ -4,7 +4,7 @@ version: 1.0.0
 homepage: https://github.com/Sub6Resources/flutter_html
 
 environment:
-  sdk: '>=2.2.2 <4.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=1.17.0'
 
 dependencies:
@@ -23,6 +23,7 @@ dependencies:
   webview_media:
     git:
       url: https://github.com/skore-io/webview_media
+      ref: update/migrate_to_nullsafety
 
   # Plugins for rendering the <audio> tag.
   chewie_audio: ^1.3.0
@@ -35,6 +36,7 @@ dependencies:
 
   flutter:
     sdk: flutter
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   flutter_test:
