@@ -5,15 +5,15 @@ import 'package:html/dom.dart' as dom;
 
 /// An [InteractableElement] is a [StyledElement] that takes user gestures (e.g. tap).
 class InteractableElement extends StyledElement {
-  String href;
+  String? href;
 
   InteractableElement({
-    String name,
-    List<StyledElement> children,
-    Style style,
+    String? name,
+    List<StyledElement>? children,
+    Style? style,
     this.href,
-    dom.Node node,
-  }) : super(name: name, children: children, style: style, node: node);
+    dom.Node? node,
+  }) : super(name: name, children: children, style: style, node: node as dom.Element?);
 }
 
 /// A [Gesture] indicates the type of interaction by a user.
